all tests are now correctly organized by CONFIG_VFS_SUPPORT_DIR
